requirements.txt
# DEPENDENCIAS

git+https://github.com/Pycord-Development/pycord
python-dotenv
requests
pandas
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib